Proyecto3/fx-service/requirements.txt
# requirements.txt — FX Service (DeliverEats)
Flask==3.0.0
flask-cors==4.0.0
grpcio==1.60.0
grpcio-tools==1.60.0
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
pytest==8.0.0
pytest-mock==3.12.0
pytest-cov==4.1.0
coverage==7.4.0
